[UI]: contact fragment
- add collapsing tollbar with main logo
