+++ app/src/main/java/com/android/contact_119/MainActivity.kt
@@ -2,10 +2,9 @@ package com.android.contact_119
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.widget.TextView
+import com.android.contact_119.adapter.ViewPagerAdapter
 import com.android.contact_119.databinding.ActivityMainBinding
 import com.google.android.material.tabs.TabLayoutMediator
-import java.time.format.TextStyle
 
 class MainActivity : AppCompatActivity() {
     private lateinit var binding: ActivityMainBinding
