+++ app/src/main/java/com/android/contact_119/fragment/ContactFragment.kt
@@ -1,0 +1,25 @@
+package com.android.contact_119.fragment
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.android.contact_119.R
+import com.android.contact_119.databinding.FragmentContactBinding
+
+class ContactFragment : Fragment() {
+    private val binding by lazy { FragmentContactBinding.inflate(layoutInflater) }
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+    }
+
+}
